[Fix] 소수점거래 quantity float 수정

+++ trading-service/src/main/java/com/example/trading_service/service/TradingService.java
@@ -442,11 +442,11 @@ public class TradingService {
         if (order.getOrderType() == Order.OrderType.BUY) {
             // 매수: 잔고 차감, 보유 종목 추가/업데이트
             updateBalance(order.getInvestmentAccount().getInvestmentAccountId(), -totalAmount);
-            updateHolding(order.getInvestmentAccount().getInvestmentAccountId(), order.getStock().getId(), (int) order.getQuantity(), executionPrice, true);
+            updateHolding(order.getInvestmentAccount().getInvestmentAccountId(), order.getStock().getId(), order.getQuantity(), executionPrice, true);
         } else {
             // 매도: 잔고 증가, 보유 종목 차감
             updateBalance(order.getInvestmentAccount().getInvestmentAccountId(), totalAmount);
-            updateHolding(order.getInvestmentAccount().getInvestmentAccountId(), order.getStock().getId(), (int) order.getQuantity(), executionPrice, false);
+            updateHolding(order.getInvestmentAccount().getInvestmentAccountId(), order.getStock().getId(), order.getQuantity(), executionPrice, false);
         }
     }
